refactor(battery): remove legacy properties
