+++ docs/docs/segment-battery.md
@@ -17,18 +17,17 @@ Battery displays the remaining power percentage for your battery.
   "powerline_symbol": "\uE0B0",
   "foreground": "#193549",
   "background": "#ffeb3b",
+  "background_templates": [
+    "{{if eq \"Charging\" .State.String}}#40c4ff{{end}}",
+    "{{if eq \"Discharging\" .State.String}}#ff5722{{end}}",
+    "{{if eq \"Full\" .State.String}}#4caf50{{end}}"
+  ],
   "properties": {
-    "battery_icon": "",
     "discharging_icon": "\uE231 ",
     "charging_icon": "\uE234 ",
     "charged_icon": "\uE22F ",
-    "color_background": true,
-    "charged_color": "#4caf50",
-    "charging_color": "#40c4ff",
-    "discharging_color": "#ff5722",
     "postfix": "\uF295 ",
-    "display_charging": true,
-    "display_charged": true
+    "template": "{{ if not .Error }}{{.Icon}}{{.Percentage}}{{ end }}"
   }
 }
 ```
@@ -41,16 +40,19 @@ properties below - defaults to `{{.Icon}}{{ if not .Error }}{{.Percentage}}{{ en
 - charging_icon: `string` - icon to display on the left when charging - defaults to empty
 - discharging_icon: `string` - icon to display on the left when discharging - defaults to empty
 - charged_icon: `string` - icon to display on the left when fully charged - defaults to empty
-- color_background: `boolean` - color the background or foreground for properties below - defaults to `false`
-- charged_color: `string` [color][colors] - color to use when fully charged - defaults to segment color
-- charging_color: `string` [color][colors] - color to use when charging - defaults to segment color
-- discharging_color: `string` [color][colors] - color to use when discharging - defaults to segment color
-- display_charging: `bool` - displays the battery status while charging (Charging)
-- display_charged: `bool` - displays the battery status when charged (Full)
 
 ## Template Properties
 
-- `.Battery`: `struct` - the [battery][battery] object, you can use any property it has e.g. `.Battery.State`
+- `.State`: `struct` - the battery state, has a `.String` function
+- `.Current`: `float64` - Current (momentary) charge rate (in mW).
+- `.Full`: `float64` - Last known full capacity (in mWh)
+- `.Design`: `float64` - Reported design capacity (in mWh)
+- `.ChargeRate`: `float64` - Current (momentary) charge rate (in mW). It is always non-negative, consult .State
+field to check whether it means charging or discharging
+- `.Voltage`: `float64` - Current voltage (in V)
+- `.DesignVoltage`: `float64` - Design voltage (in V). Some systems (e.g. macOS) do not provide a separate
+value for this. In such cases, or if getting this fails, but getting `Voltage` succeeds, this field will have
+the same value as `Voltage`, for convenience
 - `.Percentage`: `float64` - the current battery percentage
 - `.Error`: `string` - the error in case fetching the battery information failed
 - `.Icon`: `string` - the icon based on the battery state
