+++ themes/atomicBit.omp.json
@@ -182,14 +182,16 @@
         {
           "type": "battery",
           "style": "plain",
+          "foreground": "#ffffff",
+          "foreground_templates": [
+            "{{if eq \"Charging\" .State.String}}#40c4ff{{end}}",
+            "{{if eq \"Discharging\" .State.String}}#FFFB38{{end}}",
+            "{{if eq \"Full\" .State.String}}#33DD2D{{end}}"
+          ],
           "properties": {
             "discharging_icon": " ",
             "charging_icon": "\uf1e6 ",
             "charged_icon": "\uf58e ",
-            "color_foreground": true,
-            "charged_color": "#33DD2D",
-            "charging_color": "#40c4ff",
-            "discharging_color": "#FFFB38",
             "prefix": "<#ffffff>[</>",
             "postfix": "%<#ffffff>]─</>"
           }
