+++ themes/bubblesline.omp.json
@@ -117,18 +117,17 @@
           "leading_diamond": " \uE0B6",
           "foreground": "#9B6BDF",
           "background": "#424242",
+          "foreground_templates": [
+            "{{if eq \"Charging\" .State.String}}#40c4ff{{end}}",
+            "{{if eq \"Discharging\" .State.String}}#ff5722{{end}}",
+            "{{if eq \"Full\" .State.String}}#4caf50{{end}}"
+          ],
           "properties": {
-            "battery_icon": "",
             "discharging_icon": "\u21E3 ",
             "charging_icon": "\u21E1 ",
             "charged_icon": "\u25CF ",
-            "color_background": false,
-            "charged_color": "#4caf50",
-            "charging_color": "#40c4ff",
-            "discharging_color": "#ff5722",
             "prefix": "",
-            "postfix": "",
-            "display_charging": true
+            "postfix": ""
           }
         }
       ]
