+++ themes/negligible.omp.json
@@ -78,12 +78,14 @@
           "type": "battery",
           "style": "powerline",
           "invert_powerline": true,
+          "foreground_templates": [
+            "{{if eq \"Charging\" .State.String}}#4caf50{{end}}",
+            "{{if eq \"Discharging\" .State.String}}#40c4ff{{end}}",
+            "{{if eq \"Full\" .State.String}}#ff0000{{end}}"
+          ],
           "properties": {
             "charging_icon": " ",
             "charged_icon": "\uf00d ",
-            "charged_color": "#ff0000",
-            "charging_color": "#4caf50",
-            "discharging_color": "#40c4ff",
             "prefix": "| ",
             "postfix": "  "
           }
